[#263] Simplify 'Codec' abstraction
Resolves #263
