+++ tomland.cabal
@@ -1,6 +1,6 @@
 cabal-version:       2.4
 name:                tomland
-version:             1.2.1.0
+version:             1.3.0.0
 synopsis:            Bidirectional TOML serialization
 description:
     Implementation of bidirectional TOML serialization. Simple codecs look like this:
@@ -80,10 +80,13 @@ library
 
   exposed-modules:     Toml
                          Toml.Bi
-                           Toml.Bi.Code
-                           Toml.Bi.Combinators
-                           Toml.Bi.Monad
                            Toml.Bi.Map
+                         Toml.Codec
+                           Toml.Codec.Code
+                           Toml.Codec.Combinators
+                           Toml.Codec.Di
+                           Toml.Codec.Error
+                           Toml.Codec.Types
                          Toml.Edsl
                          Toml.Generic
                          Toml.Parser
